fix(cluster): rebase onto the merged fixes, and split the driver's record types
Rebasing #126 onto a main that now carries #130 and #131 put two independently
added tests at the same point in teardown.rs. My first resolution kept "both
sides" of a conflict whose hunks were mid-function, which spliced two test
bodies together into something that did not parse -- caught immediately by
cargo, and redone by taking main's file and re-applying the one test on top.

The combined file then crossed the 500-line cap. The obvious seam -- supervision
and stop_cluster, which share the same honesty problem -- cannot be split out: a
trait impl may not span files, and both are `ClusterControl` methods. So the
record types (`Target`, `Torn`, `Running`, `Pending`) move to
`clusterdriver/types.rs` instead, and four comments I had written at three times
the necessary length are cut to what they need to say.

No behaviour change in this commit: 371 agent tests and 149 atlasctl tests pass,
clippy and fmt clean, nothing over the cap.
